edgeNode/requirements.txt
paho-mqtt
pycryptodome
psutil